worker/requirements.txt
runpod
torch>=2.2.0
transformers>=4.49.0
Pillow
timm
opencv-python-headless
numpy
